Fix .last function
Added return statement that returns the last element if n is undefined by using array.length - 1 for the index. If n is passed, I used .slice(Math.max(array.length - n, 0)). The .slice method will take a portion of the array and pass back a new array with given indexes. I used Math.max to ensure that numbers >= 0 are passed. This is helpful for when n is greater than our length; the end results becomes .slice(0), which returns the entire array.