Add index and collection to parameters to iterator function in .each
Needed to be able to iterate over each element in the collection while providing access to the index and the original collection. To provide access to the index, I simply set the index parameter to i, which we're using to iterate. To provide access to the original collection, I set the list parameter to collection, which will be passed in .each argument.