Create _.reject function
Used the _.filter function to find an array that pass a truth test. I then checked to see if the collection had any values that were included in that array, and then pushed the ones that weren't included. I'm not sure if there is a better way to implement the _.filter function here. I feel like simply changing the code from the _.filter function to check for non-passing elements would be faster. I can maybe revisit.