Create reduce function
Created logic for when accumulator is defined and undefined:

For case 1, I used a simple for loop, changing the value of accumulator at the end of each loop to be the result of the iterator function calling on accumulator.

For case 2, I used the same for loop strategy, but changed the initial value of accumulator to be the first element of collection and started the loop at the 2nd value of collection.