Create _.every function
Used the _.reduce function to check if each value is true. If the first value is true, it will return the accumulator which was initially defined as true. Once a value is false, it will return false, thus changing the value of the accumulator to false and then passing false for every value after.